+++ apps/pizza_api/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Some small documentation
+- Tests config
 
 ## [0.5.1] - 2026-08-03