+++ apps/pizza_app/CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Some documentation and quality of tooling
+- Tests config
 
 ## [0.1.2] - 2026-07-04